Add a LogListener option to Notepad
For clients using the "log counting" feature, this will be slightly breaking. But it adds lots of flexibility which should make it worth while.

Now you can supply one or more `LogListener` funcs when creating a new `Notepad`. Typical use cases would be to capture ERROR logs in unit tests, and to count log statements. A listener for log counting is provided, see the `LogCounter` func.

If you want to add listeners to the global logger, see `SetLogListeners`.